# fix: enoding partial paths for leaf nodes