@demosdemon 2025-07-16
> This looks great, did you run any of the serialization perf tests to verify there's no drop in performance?

good call on perf check. There was a slight regression (<200ns). I'm inclined to attribute that to the compiler no longer eliding the branch for the leaf node. Switching to a cold fn didn't help much.

I got some time back by not filling the vecs with zeros when we allocate them, but it didn't give me back much (<100 ns).