@demosdemon 2025-07-16
current perf. previously I ran benchmarks with release and not maxperf. release still shows a regression but with maxperf, it's equivalent.

### before (`maxperf`)

```text
     Running benches/serializer.rs (target/maxperf/deps/serializer-7dba99e4cfe6e226)
leaf/manual             time:   [17.795 ns 17.856 ns 17.925 ns]
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

has_value/manual        time:   [24.190 ns 24.306 ns 24.420 ns]

1_child/manual          time:   [21.260 ns 21.354 ns 21.453 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

2_child/manual          time:   [22.335 ns 22.397 ns 22.456 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high severe

16_child/manual         time:   [29.316 ns 29.420 ns 29.534 ns]
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe
```

### after (`maxperf`)

```text
     Running benches/serializer.rs (target/maxperf/deps/serializer-7dba99e4cfe6e226)
leaf/manual             time:   [17.991 ns 18.046 ns 18.107 ns]
                        change: [+0.9315% +1.6199% +2.4299%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 8 outliers among 100 measurements (8.00%)
  6 (6.00%) high mild
  2 (2.00%) high severe

has_value/manual        time:   [23.875 ns 23.978 ns 24.091 ns]
                        change: [−1.1993% −0.6307% −0.0375%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

1_child/manual          time:   [21.035 ns 21.092 ns 21.152 ns]
                        change: [−0.7356% −0.2167% +0.2966%] (p = 0.44 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  1 (1.00%) low mild
  4 (4.00%) high mild
  4 (4.00%) high severe

2_child/manual          time:   [22.139 ns 22.246 ns 22.369 ns]
                        change: [−1.1173% −0.4746% +0.0942%] (p = 0.13 > 0.05)
                        No change in performance detected.

16_child/manual         time:   [28.591 ns 28.693 ns 28.792 ns]
                        change: [−3.5903% −3.0764% −2.6146%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
```

### before (`release`)

```text
     Running benches/serializer.rs (target/release/deps/serializer-a4a815a6490fbc6e)
leaf/manual             time:   [23.328 ns 23.399 ns 23.471 ns]
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) high mild
  1 (1.00%) high severe

has_value/manual        time:   [35.302 ns 35.488 ns 35.714 ns]
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) high mild
  1 (1.00%) high severe

1_child/manual          time:   [31.525 ns 31.652 ns 31.784 ns]
Found 7 outliers among 100 measurements (7.00%)
  6 (6.00%) high mild
  1 (1.00%) high severe

2_child/manual          time:   [39.008 ns 39.234 ns 39.498 ns]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

16_child/manual         time:   [95.326 ns 95.600 ns 95.904 ns]
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
```

### after (`release`)

```text
     Running benches/serializer.rs (target/release/deps/serializer-a4a815a6490fbc6e)
leaf/manual             time:   [24.409 ns 24.574 ns 24.751 ns]
                        change: [+4.8989% +5.6708% +6.5487%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 4 outliers among 100 measurements (4.00%)
  3 (3.00%) high mild
  1 (1.00%) high severe

has_value/manual        time:   [36.525 ns 36.764 ns 37.020 ns]
                        change: [+4.3474% +5.1041% +5.9831%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

1_child/manual          time:   [32.614 ns 32.796 ns 32.969 ns]
                        change: [+1.8994% +2.4836% +3.1077%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

2_child/manual          time:   [40.387 ns 40.578 ns 40.775 ns]
                        change: [+2.6220% +3.2309% +3.8581%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

16_child/manual         time:   [95.056 ns 95.470 ns 96.012 ns]
                        change: [−0.9768% −0.3449% +0.3054%] (p = 0.29 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe
```