@demosdemon 2025-07-16
### before (release)

```text
     Running benches/serializer.rs (target/maxperf/deps/serializer-72b5b75c432c3437)
leaf/manual             time:   [19.503 ns 19.717 ns 19.963 ns]
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
leaf/from_reader        time:   [43.607 ns 43.846 ns 44.087 ns]
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) low mild
  5 (5.00%) high mild

has_value/manual        time:   [25.695 ns 25.818 ns 25.939 ns]
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) high mild
  3 (3.00%) high severe
has_value/from_reader   time:   [200.05 ns 201.08 ns 202.16 ns]
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild

1_child/manual          time:   [22.668 ns 22.813 ns 22.930 ns]
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
1_child/from_reader     time:   [168.14 ns 169.66 ns 171.27 ns]
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild

2_child/manual          time:   [24.087 ns 24.387 ns 24.749 ns]
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe
2_child/from_reader     time:   [242.71 ns 244.28 ns 246.05 ns]
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild

16_child/manual         time:   [30.195 ns 30.370 ns 30.576 ns]
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
16_child/from_reader    time:   [1.2691 µs 1.2755 µs 1.2828 µs]
Found 7 outliers among 100 measurements (7.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
  4 (4.00%) high severe
```

### after (release)

```text
     Running benches/serializer.rs (target/maxperf/deps/serializer-72b5b75c432c3437)
leaf/manual             time:   [19.232 ns 19.311 ns 19.401 ns]
                        change: [−3.4699% −2.5088% −1.5599%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 7 outliers among 100 measurements (7.00%)
  4 (4.00%) high mild
  3 (3.00%) high severe
leaf/from_reader        time:   [40.378 ns 40.477 ns 40.585 ns]
                        change: [−6.9701% −6.3303% −5.6740%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  5 (5.00%) high mild
  1 (1.00%) high severe

has_value/manual        time:   [25.446 ns 25.535 ns 25.638 ns]
                        change: [−2.5739% −1.8665% −1.1303%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  4 (4.00%) high mild
  2 (2.00%) high severe
has_value/from_reader   time:   [197.96 ns 198.85 ns 199.73 ns]
                        change: [−2.0173% −1.4189% −0.8347%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

1_child/manual          time:   [22.367 ns 22.426 ns 22.484 ns]
                        change: [−2.0640% −1.5789% −1.0891%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
1_child/from_reader     time:   [196.26 ns 197.00 ns 197.80 ns]
                        change: [+14.421% +15.660% +16.858%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) low severe
  3 (3.00%) low mild
  2 (2.00%) high severe

2_child/manual          time:   [23.416 ns 23.485 ns 23.552 ns]
                        change: [−3.3504% −2.4636% −1.6114%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  1 (1.00%) low mild
  2 (2.00%) high mild
2_child/from_reader     time:   [270.75 ns 272.41 ns 274.00 ns]
                        change: [+8.9115% +9.9633% +11.013%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 5 outliers among 100 measurements (5.00%)
  4 (4.00%) low mild
  1 (1.00%) high mild

16_child/manual         time:   [30.159 ns 30.214 ns 30.276 ns]
                        change: [−0.9819% −0.1989% +0.4029%] (p = 0.63 > 0.05)
                        No change in performance detected.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) high mild
  2 (2.00%) high severe
16_child/from_reader    time:   [1.2840 µs 1.2902 µs 1.2966 µs]
                        change: [+0.0007% +1.1588% +2.0984%] (p = 0.03 < 0.05)
                        Change within noise threshold.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) low mild
  2 (2.00%) high mild
```